feat(comparateur): remet les cartes de statut
